backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
watchdog>=3.0.0
apscheduler>=3.10.0
aiofiles>=23.0.0
httpx>=0.25.0
